# docs/swagger.yaml
info:
  title: ALSIRAJ NODEJS SWAGGER  DOCUMENTATION
  version: 1.0.0
  description: __*This describes services details of ALSIRAJ*__
  termsOfService: 'https://alsiraj-website/terms-conditions/overview'
  contact:
    email: 'suriyathangaraman@gmail.com'
  license:
    name: ''
    url: 'https://www.alsiraj.com'
basePath: '/'
swagger: '2.0'
...
schemes:
  - http
  - https
securityDefinitions:
    Bearer:
        type: apiKey
        description: "Value: Bearer "
        name: Authorization
        in: header
security:
  - Bearer: []
paths:
...
  /masters/any/competition/add:
    post:
      tags:
        - 'Competitions'
      summary: To add new competitions
      operationId: _addEvent
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/EventInputInformation'
      responses:
        '200':
          description: OK
          schema:
            type: 'object'
            properties:
              data:
                $ref: '#/definitions/GenralResponse'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: To create new competition
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
  /masters/any/competition/edit/{id}:
    put:
      tags:
        - 'Competitions'
      summary: To update existing competition details
      operationId: _editEvent
      parameters:
       - in: path
         name: id
         required: true
         type: string
       - in: body
         name: body
         required: true
         schema:
            $ref: '#/definitions/EventInputInformation'
      responses:
        '200':
          description: OK
          schema:
            type: 'object'
            properties:
              data:
                $ref: '#/definitions/GenralResponse'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description:  To udpate new competition
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
  /masters/any/competition/delete/{id}:
    delete:
      tags:
        - 'Competitions'
      summary: To delete existing competition
      operationId: _deleteEvent
      parameters:
       - in: path
         name: id
         required: true
         type: string
      responses:
        '200':
          description: OK
          schema:
            type: 'object'
            properties:
              data:
                $ref: '#/definitions/GenralResponse'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: To delete new competition
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
  /masters/any/competition/list:
    get:
      tags:
        - 'Competitions'
      summary: To get existing competitions
      operationId: _getEvents
      responses:
        '200':
          description: OK
          schema:
            type: 'object'
            properties:
              data:
                type: 'array'
                items:
                 $ref: '#/definitions/GetEvents'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: To get existing competitions
      consumes:
        - 'application/json'
      produces:
        - 'application/json'      
...
definitions:
...
  EventInputInformation:
    type: object
    properties:
      competitionName:
        type: string
      type:
        type: string
      status:
        type: string
      startDate:
        type: string
      endDate:
        type: string
      description:
        type: string
      documentUrl:
        type: string  
...
  GetEvents:
    type: object
    properties:
      id:
        type: string
      competitionName:
        type: string
      type:
        type: string
      status:
        type: string
      startDate:
        type: string
      endDate:
        type: string
      description:
        type: number        
  GenralResponse:
    type: object
    properties:
     status:
      type: number
     message:
      type: string
...
  ErrorResponse:
    type: object
    properties:
      httpErrorMessage:
        type: 'string'
      message:
        type: 'string'              